Better enum logic

+++ MoveResult.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 namespace ChessersEngine {
     public class MoveAttempt {
-        public long playerId { get; set; }
+        public int playerId { get; set; }
         public int pieceGuid { get; set; }
         public int pieceId { get; set; }
         public int tileId { get; set; }
@@ -17,7 +17,7 @@ namespace ChessersEngine {
     }
 
     public class MoveResult {
-        public long playerId { get; set; }
+        public int playerId { get; set; }
 
         public int pieceId { get; set; }
         public int pieceGuid { get; set; }
